Domains: Remove redux-bridge from domain management
